+++ BitFaster.Caching/Lru/ConcurrentLruCore.cs
@@ -88,7 +88,7 @@ namespace BitFaster.Caching.Lru
             this.warmQueue = new ConcurrentQueue<I>();
             this.coldQueue = new ConcurrentQueue<I>();
 
-            int dictionaryCapacity = HashTableSize.Estimate(this.Capacity);
+            int dictionaryCapacity = ConcurrentDictionarySize.Estimate(this.Capacity);
 
             this.dictionary = new ConcurrentDictionary<K, I>(concurrencyLevel, dictionaryCapacity, comparer);
             this.itemPolicy = itemPolicy;
